# Fix nextjs phaser issue