--- a/packages/frontend/package.json
+++ b/packages/frontend/package.json
@@ -22,7 +22,7 @@
     "lodash-es": "^4.17.21",
     "lucide-react": "^0.501.0",
     "motion": "^12.7.4",
-    "next": "~15.1.7",
+    "next": "~15.3.1",
     "phaser": "^3.88.2",
     "react": "^19.1.0",
     "react-confetti": "^6.4.0",