--- a/packages/frontend/src/app/[gameSlug]/[gameId]/global/page.tsx
+++ b/packages/frontend/src/app/[gameSlug]/[gameId]/global/page.tsx
@@ -1,10 +1,15 @@
 "use client";
 
-import { GlobalScreen } from "@/components/globalScreen/GlobalScreen";
 import { GlobalScreenPlayerContextProvider } from "@/components/globalScreen/GlobalScreenPlayerContextProvider";
 import { ClientGate } from "@/lib/ClientGate";
 import { GameId, GameType } from "@resync-games/api";
+import dynamic from "next/dynamic";
 import { useParams } from "next/navigation";
+
+const GlobalScreenWithoutSSR = dynamic(
+  () => import("@/components/globalScreen/GlobalScreen"),
+  { ssr: false }
+);
 
 export default function Page() {
   const { gameSlug, gameId } = useParams<{
@@ -15,7 +20,7 @@ export default function Page() {
   return (
     <ClientGate>
       <GlobalScreenPlayerContextProvider gameId={gameId}>
-        <GlobalScreen gameId={gameId} gameSlug={gameSlug} />
+        <GlobalScreenWithoutSSR gameId={gameId} gameSlug={gameSlug} />
       </GlobalScreenPlayerContextProvider>
     </ClientGate>
   );