--- a/packages/frontend/src/components/player/PlayerContext.tsx
+++ b/packages/frontend/src/components/player/PlayerContext.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { useNetworkCall } from "@/lib/hooks/useNetworkCall";
 import { ClientServiceCallers } from "@/services/serviceCallers";
 import { GameId, Player } from "@resync-games/api";