- Fixes some SSR issues
- Upgrades NextJS
- Sorts stock in the stock times by order
- Increases cycle time for stocks

---

This pull request introduces several updates to the frontend and game components, including improvements to server-side rendering (SSR), dependency updates, and code simplifications. The most significant changes involve enabling dynamic imports for specific components to disable SSR, updating the `next` package version, and optimizing selectors and state management in the "The Stock Times" game.

### Frontend Changes

* **Server-Side Rendering (SSR) Adjustments**:
  - Replaced `GlobalScreen` and `GetGameState` components with dynamically imported versions (`GlobalScreenWithoutSSR` and `GetGameStateWithoutSSR`) to disable SSR for these components. (`[packages/frontend/src/app/[gameSlug]/[gameId]/global/page.tsxL3-R13](diffhunk://#diff-8df7b9f91e740448042cab9ccf2c8b768930b6801d87ffe9ffd5a8a00f046c06L3-R13)`, `[packages/frontend/src/app/[gameSlug]/[gameId]/global/page.tsxL18-R23](diffhunk://#diff-8df7b9f91e740448042cab9ccf2c8b768930b6801d87ffe9ffd5a8a00f046c06L18-R23)`, `[packages/frontend/src/app/[gameSlug]/[gameId]/page.tsxL3-R13](diffhunk://#diff-c072cf45eef77c28e7a54c3aae8c6548890e47841fb908c47b983804beb81d9cL3-R13)`, `[packages/frontend/src/app/[gameSlug]/[gameId]/page.tsxL18-R23](diffhunk://#diff-c072cf45eef77c28e7a54c3aae8c6548890e47841fb908c47b983804beb81d9cL18-R23)`)
  - Added a fallback check in `ClientGate` to return `null` when `window` is undefined, improving SSR compatibility. (`[packages/frontend/src/lib/ClientGate.tsxR10-R13](diffhunk://#diff-efbb4461dc879eb430ebbee4218048d7d93b23e0fa1fb014b3b706b36ae3dc56R10-R13)`)

* **Dependency Updates**:
  - Updated the `next` package from version `~15.1.7` to `~15.3.1` in `package.json`. (`[packages/frontend/package.jsonL25-R25](diffhunk://#diff-8571ecd91584b00015b23695d3a6a164282636bb47bfbe46dca243bf9b4db773L25-R25)`)

### Game-Specific Changes

* **Selector Optimization in "The Stock Times"**:
  - Introduced a new selector, `selectStocksWithOrder`, to retrieve stocks with their order index, replacing manual sorting logic in `CorruptStock` and `StockLock` components. (`[[1]](diffhunk://#diff-dcc930a3e1cd3ef03fb28bf5ccf293724403a2338765212f13ca02e05dc6e3e0R196-R223)`, `[[2]](diffhunk://#diff-4d2f4d08677688b61893dda10c9c757f381cf122065900ef5d9b87e3fbc3ca31L24-R29)`, `[[3]](diffhunk://#diff-4d2f4d08677688b61893dda10c9c757f381cf122065900ef5d9b87e3fbc3ca31L100-R96)`, `[[4]](diffhunk://#diff-ebe5ad47f9e0d2b39fb28618412a7caf63b6231a0dcc0c5fec6557ac48be0054L27-R34)`, `[[5]](diffhunk://#diff-ebe5ad47f9e0d2b39fb28618412a7caf63b6231a0dcc0c5fec6557ac48be0054L117-R114)`)

* **Configuration Update**:
  - Adjusted the default stock cycle time from 12 seconds to 15 seconds and increased the maximum value from 21 to 30 seconds in `TheStockTimesConfiguration`. (`[packages/games/src/frontend/theStockTimes/theStockTimesConfiguration.tsL16-R19](diffhunk://#diff-4c98e782ad836edee6ffb196ee3d2a53237c2e3f3bfc296cc38b0e609cfa8268L16-R19)`)

### Code Simplifications

* **Component Exports**:
  - Converted `GlobalScreen` and `GetGameState` components to default exports. (`[[1]](diffhunk://#diff-c348e74ff7f400c85e9e535e45ff14fb9a8d058813ba407ae10ff0449c605c64R50-R51)`, `[[2]](diffhunk://#diff-b3ff7308edbc2dc2a6e1d941f4e9d33bc35f6bea2daa8006ba2b82ef386f912eR65-R66)`)
  - Removed unused `useMemo` imports and redundant sorting logic in multiple game components. (`[[1]](diffhunk://#diff-4d2f4d08677688b61893dda10c9c757f381cf122065900ef5d9b87e3fbc3ca31L2-L10)`, `[[2]](diffhunk://#diff-ebe5ad47f9e0d2b39fb28618412a7caf63b6231a0dcc0c5fec6557ac48be0054L1-L11)`)

* **Miscellaneous**:
  - Reduced the minimum invite code length from 4 to 3 characters in the `JoinGame` component. (`[packages/frontend/src/components/joinGame/JoinGame.tsxL47-R45](diffhunk://#diff-718c32d71445041007794c67e0881693566fbcaca49cb78bd7a15db34e28324dL47-R45)`)